[DEV] Fix some code smells
